Refactor(processors): Further encapsulate policy related code.
