+++ src/main/java/com/synopsys/integration/alert/provider/blackduck/collector/builder/policy/PolicyClearedMessageBuilder.java
@@ -23,7 +23,6 @@
 package com.synopsys.integration.alert.provider.blackduck.collector.builder.policy;
 
 import java.util.ArrayList;
-import java.util.Collection;
 import java.util.Date;
 import java.util.LinkedList;
 import java.util.List;
@@ -33,7 +32,6 @@ import java.util.Set;
 import java.util.function.Function;
 import java.util.stream.Collectors;
 
-import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -52,8 +50,6 @@ import com.synopsys.integration.alert.provider.blackduck.collector.builder.Messa
 import com.synopsys.integration.alert.provider.blackduck.collector.builder.util.ComponentBuilderUtil;
 import com.synopsys.integration.alert.provider.blackduck.collector.builder.util.PolicyPriorityUtil;
 import com.synopsys.integration.alert.provider.blackduck.collector.util.BlackDuckResponseCache;
-import com.synopsys.integration.blackduck.api.generated.component.PolicyRuleExpressionSetView;
-import com.synopsys.integration.blackduck.api.generated.component.PolicyRuleExpressionView;
 import com.synopsys.integration.blackduck.api.generated.enumeration.NotificationType;
 import com.synopsys.integration.blackduck.api.generated.view.PolicyRuleView;
 import com.synopsys.integration.blackduck.api.generated.view.VersionBomComponentView;
@@ -72,11 +68,13 @@ public class PolicyClearedMessageBuilder implements BlackDuckMessageBuilder<Rule
     private final Logger logger = LoggerFactory.getLogger(PolicyClearedMessageBuilder.class);
     private PolicyPriorityUtil policyPriorityUtil;
     private ComponentBuilderUtil componentBuilderUtil;
+    private PolicyCommonBuilder policyCommonBuilder;
 
     @Autowired
-    public PolicyClearedMessageBuilder(PolicyPriorityUtil policyPriorityUtil, ComponentBuilderUtil componentBuilderUtil) {
+    public PolicyClearedMessageBuilder(PolicyPriorityUtil policyPriorityUtil, ComponentBuilderUtil componentBuilderUtil, PolicyCommonBuilder policyCommonBuilder) {
         this.policyPriorityUtil = policyPriorityUtil;
         this.componentBuilderUtil = componentBuilderUtil;
+        this.policyCommonBuilder = policyCommonBuilder;
     }
 
     @Override
@@ -101,7 +99,7 @@ public class PolicyClearedMessageBuilder implements BlackDuckMessageBuilder<Rule
                                                                               .applySubTopic(MessageBuilderConstants.LABEL_PROJECT_VERSION_NAME, violationContent.getProjectVersionName(), violationContent.getProjectVersion())
                                                                               .applyProviderCreationTime(providerCreationDate);
             Map<String, PolicyInfo> policyUrlToInfoMap = violationContent.getPolicyInfos().stream().collect(Collectors.toMap(PolicyInfo::getPolicy, Function.identity()));
-            SetMap<ComponentVersionStatus, PolicyInfo> componentPolicies = createComponentToPolicyMapping(violationContent.getComponentVersionStatuses(), policyUrlToInfoMap);
+            SetMap<ComponentVersionStatus, PolicyInfo> componentPolicies = policyCommonBuilder.createComponentToPolicyMapping(violationContent.getComponentVersionStatuses(), policyUrlToInfoMap);
             List<ComponentItem> items = new LinkedList<>();
             for (Map.Entry<ComponentVersionStatus, Set<PolicyInfo>> componentToPolicyEntry : componentPolicies.entrySet()) {
                 ComponentVersionStatus componentVersionStatus = componentToPolicyEntry.getKey();
@@ -132,8 +130,8 @@ public class PolicyClearedMessageBuilder implements BlackDuckMessageBuilder<Rule
             Optional<VersionBomComponentView> optionalBomComponent = blackDuckResponseCache.getBomComponentView(bomComponentUrl);
 
             List<LinkableItem> policyAttributes = new ArrayList<>();
-            LinkableItem policyNameItem = createPolicyNameItem(policyInfo);
-            LinkableItem nullablePolicySeverityItem = createPolicySeverityItem(policyInfo).orElse(null);
+            LinkableItem policyNameItem = componentBuilderUtil.createPolicyNameItem(policyInfo);
+            LinkableItem nullablePolicySeverityItem = componentBuilderUtil.createPolicySeverityItem(policyInfo).orElse(null);
             optionalBomComponent.ifPresent(bomComponent -> {
                 policyAttributes.addAll(componentBuilderUtil.getLicenseLinkableItems(bomComponent));
                 policyAttributes.addAll(componentBuilderUtil.getUsageLinkableItems(bomComponent));
@@ -157,68 +155,13 @@ public class PolicyClearedMessageBuilder implements BlackDuckMessageBuilder<Rule
                 logger.info("Error building policy component for notification {}, operation {}, component {}, component version {}", notificationId, operation, componentName, componentVersionName);
                 logger.error("Error building policy component cause ", ex);
             }
-            Optional<PolicyRuleView> optionalPolicyRule = getPolicyRule(blackDuckResponseCache, policyInfo);
-            if (optionalPolicyRule.isPresent() && hasVulnerabilityRule(optionalPolicyRule.get())) {
+            Optional<PolicyRuleView> optionalPolicyRule = blackDuckResponseCache.getPolicyRule(blackDuckResponseCache, policyInfo);
+            if (optionalPolicyRule.isPresent() && policyCommonBuilder.hasVulnerabilityRule(optionalPolicyRule.get())) {
                 Optional<ComponentItem> vulnerabilityComponent = createEmptyVulnerabilityItem(blackDuckResponseCache, policyNameItem, componentName, componentVersionName, projectVersionUrl, notificationId, operation);
                 vulnerabilityComponent.ifPresent(componentItems::add);
             }
         }
         return componentItems;
-    }
-
-    protected LinkableItem createPolicyNameItem(PolicyInfo policyInfo) {
-        String policyName = policyInfo.getPolicyName();
-        return new LinkableItem(MessageBuilderConstants.LABEL_POLICY_NAME, policyName);
-    }
-
-    protected Optional<LinkableItem> createPolicySeverityItem(PolicyInfo policyInfo) {
-        String severity = policyInfo.getSeverity();
-        if (StringUtils.isNotBlank(severity)) {
-            LinkableItem severityItem = new LinkableItem(MessageBuilderConstants.LABEL_POLICY_SEVERITY_NAME, severity);
-            return Optional.of(severityItem);
-        }
-        return Optional.empty();
-    }
-
-    public boolean hasVulnerabilityRule(PolicyRuleView policyRule) {
-        PolicyRuleExpressionSetView expression = policyRule.getExpression();
-        List<PolicyRuleExpressionView> expressions = expression.getExpressions();
-        for (PolicyRuleExpressionView expressionView : expressions) {
-            if (expressionView.getName().toLowerCase().contains(MessageBuilderConstants.VULNERABILITY_CHECK_TEXT)) {
-                return true;
-            }
-        }
-        return false;
-    }
-
-    public Optional<PolicyRuleView> getPolicyRule(BlackDuckResponseCache blackDuckResponseCache, PolicyInfo policyInfo) {
-        try {
-            String policyUrl = policyInfo.getPolicy();
-            if (StringUtils.isNotBlank(policyUrl)) {
-                return blackDuckResponseCache.getItem(PolicyRuleView.class, policyUrl);
-            }
-        } catch (Exception e) {
-            logger.debug("Unable to get policy rule: {}", policyInfo.getPolicyName());
-            logger.debug("Cause:", e);
-        }
-        return Optional.empty();
-    }
-
-    private SetMap<ComponentVersionStatus, PolicyInfo> createComponentToPolicyMapping(
-        Collection<ComponentVersionStatus> componentVersionStatuses, Map<String, PolicyInfo> policyItems) {
-        SetMap<ComponentVersionStatus, PolicyInfo> componentToPolicyMapping = SetMap.createDefault();
-        for (ComponentVersionStatus componentVersionStatus : componentVersionStatuses) {
-            Set<PolicyInfo> componentPolicies = getPoliciesForComponent(componentVersionStatus, policyItems);
-            componentToPolicyMapping.addAll(componentVersionStatus, componentPolicies);
-        }
-        return componentToPolicyMapping;
-    }
-
-    private Set<PolicyInfo> getPoliciesForComponent(ComponentVersionStatus componentVersionStatus, Map<String, PolicyInfo> policyItems) {
-        return componentVersionStatus.getPolicies().stream()
-                   .filter(policyItems::containsKey)
-                   .map(policyItems::get)
-                   .collect(Collectors.toSet());
     }
 
     private Optional<ComponentItem> createEmptyVulnerabilityItem(BlackDuckResponseCache blackDuckResponseCache, LinkableItem policyNameItem, String componentName, String componentVersionName, String projectVersionUrl, Long notificationId,
