+++ src/main/java/com/synopsys/integration/alert/provider/blackduck/collector/builder/policy/PolicyCommonBuilder.java
@@ -1,0 +1,204 @@
+/**
+ * blackduck-alert
+ *
+ * Copyright (c) 2019 Synopsys, Inc.
+ *
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements. See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership. The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+package com.synopsys.integration.alert.provider.blackduck.collector.builder.policy;
+
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.Comparator;
+import java.util.LinkedList;
+import java.util.List;
+import java.util.Map;
+import java.util.Optional;
+import java.util.Set;
+import java.util.stream.Collectors;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Component;
+
+import com.synopsys.integration.alert.common.SetMap;
+import com.synopsys.integration.alert.common.enumeration.ComponentItemPriority;
+import com.synopsys.integration.alert.common.enumeration.ItemOperation;
+import com.synopsys.integration.alert.common.exception.AlertException;
+import com.synopsys.integration.alert.common.message.model.ComponentItem;
+import com.synopsys.integration.alert.common.message.model.LinkableItem;
+import com.synopsys.integration.alert.provider.blackduck.collector.builder.MessageBuilderConstants;
+import com.synopsys.integration.alert.provider.blackduck.collector.builder.util.ComponentBuilderUtil;
+import com.synopsys.integration.alert.provider.blackduck.collector.builder.util.PolicyPriorityUtil;
+import com.synopsys.integration.alert.provider.blackduck.collector.builder.util.VulnerabilityUtil;
+import com.synopsys.integration.alert.provider.blackduck.collector.util.BlackDuckResponseCache;
+import com.synopsys.integration.blackduck.api.generated.component.PolicyRuleExpressionSetView;
+import com.synopsys.integration.blackduck.api.generated.component.PolicyRuleExpressionView;
+import com.synopsys.integration.blackduck.api.generated.view.PolicyRuleView;
+import com.synopsys.integration.blackduck.api.generated.view.VersionBomComponentView;
+import com.synopsys.integration.blackduck.api.generated.view.VersionBomPolicyRuleView;
+import com.synopsys.integration.blackduck.api.generated.view.VulnerabilityView;
+import com.synopsys.integration.blackduck.api.generated.view.VulnerabilityWithRemediationView;
+import com.synopsys.integration.blackduck.api.generated.view.VulnerableComponentView;
+import com.synopsys.integration.blackduck.api.manual.component.ComponentVersionStatus;
+import com.synopsys.integration.blackduck.api.manual.component.PolicyInfo;
+import com.synopsys.integration.blackduck.service.BlackDuckService;
+
+@Component
+public class PolicyCommonBuilder {
+    private final Logger logger = LoggerFactory.getLogger(PolicyCommonBuilder.class);
+    private VulnerabilityUtil vulnerabilityUtil;
+    private ComponentBuilderUtil componentBuilderUtil;
+    private PolicyPriorityUtil policyPriorityUtil;
+
+    @Autowired
+    public PolicyCommonBuilder(VulnerabilityUtil vulnerabilityUtil, ComponentBuilderUtil componentBuilderUtil, PolicyPriorityUtil policyPriorityUtil) {
+        this.vulnerabilityUtil = vulnerabilityUtil;
+        this.componentBuilderUtil = componentBuilderUtil;
+        this.policyPriorityUtil = policyPriorityUtil;
+    }
+
+    public List<ComponentItem> retrievePolicyItems(BlackDuckResponseCache blackDuckResponseCache, String componentName, String componentVersionName,
+        Collection<PolicyInfo> policies, Long notificationId, ItemOperation operation, String projectVersionUrl, String bomComponentUrl, List<LinkableItem> customAttributes) {
+        List<ComponentItem> componentItems = new LinkedList<>();
+        for (PolicyInfo policyInfo : policies) {
+            ComponentItemPriority priority = policyPriorityUtil.getPriorityFromSeverity(policyInfo.getSeverity());
+
+            Optional<VersionBomComponentView> optionalBomComponent = blackDuckResponseCache.getBomComponentView(bomComponentUrl);
+
+            List<LinkableItem> policyAttributes = new ArrayList<>();
+            LinkableItem policyNameItem = componentBuilderUtil.createPolicyNameItem(policyInfo);
+            LinkableItem nullablePolicySeverityItem = componentBuilderUtil.createPolicySeverityItem(policyInfo).orElse(null);
+            optionalBomComponent.ifPresent(bomComponent -> {
+                policyAttributes.addAll(componentBuilderUtil.getLicenseLinkableItems(bomComponent));
+                policyAttributes.addAll(componentBuilderUtil.getUsageLinkableItems(bomComponent));
+            });
+            policyAttributes.addAll(customAttributes);
+
+            try {
+                ComponentItem.Builder builder = new ComponentItem.Builder()
+                                                    .applyCategory(MessageBuilderConstants.CATEGORY_TYPE_POLICY)
+                                                    .applyOperation(operation)
+                                                    .applyPriority(priority)
+                                                    .applyCategoryItem(policyNameItem)
+                                                    .applyCategoryGroupingAttribute(nullablePolicySeverityItem)
+                                                    .applyAllComponentAttributes(policyAttributes)
+                                                    .applyNotificationId(notificationId);
+                componentBuilderUtil.applyComponentInformation(builder, blackDuckResponseCache, componentName, componentVersionName, projectVersionUrl);
+                componentItems.add(builder.build());
+            } catch (Exception ex) {
+                logger.info("Error building policy component for notification {}, operation {}, component {}, component version {}", notificationId, operation, componentName, componentVersionName);
+                logger.error("Error building policy component cause ", ex);
+            }
+        }
+        return componentItems;
+    }
+
+    public SetMap<ComponentVersionStatus, PolicyInfo> createComponentToPolicyMapping(
+        Collection<ComponentVersionStatus> componentVersionStatuses, Map<String, PolicyInfo> policyItems) {
+        SetMap<ComponentVersionStatus, PolicyInfo> componentToPolicyMapping = SetMap.createDefault();
+        for (ComponentVersionStatus componentVersionStatus : componentVersionStatuses) {
+            Set<PolicyInfo> componentPolicies = getPoliciesForComponent(componentVersionStatus, policyItems);
+            componentToPolicyMapping.addAll(componentVersionStatus, componentPolicies);
+        }
+        return componentToPolicyMapping;
+    }
+
+    public Set<PolicyInfo> getPoliciesForComponent(ComponentVersionStatus componentVersionStatus, Map<String, PolicyInfo> policyItems) {
+        return componentVersionStatus.getPolicies().stream()
+                   .filter(policyItems::containsKey)
+                   .map(policyItems::get)
+                   .collect(Collectors.toSet());
+    }
+
+    public boolean hasVulnerabilityRule(VersionBomPolicyRuleView policyRule) {
+        PolicyRuleExpressionSetView expression = policyRule.getExpression();
+        List<PolicyRuleExpressionView> expressions = expression.getExpressions();
+        for (PolicyRuleExpressionView expressionView : expressions) {
+            if (expressionView.getName().toLowerCase().contains(MessageBuilderConstants.VULNERABILITY_CHECK_TEXT)) {
+                return true;
+            }
+        }
+        return false;
+    }
+
+    public boolean hasVulnerabilityRule(PolicyRuleView policyRule) {
+        PolicyRuleExpressionSetView expression = policyRule.getExpression();
+        List<PolicyRuleExpressionView> expressions = expression.getExpressions();
+        for (PolicyRuleExpressionView expressionView : expressions) {
+            if (expressionView.getName().toLowerCase().contains(MessageBuilderConstants.VULNERABILITY_CHECK_TEXT)) {
+                return true;
+            }
+        }
+        return false;
+    }
+
+    public List<ComponentItem> createVulnerabilityPolicyComponentItems(Collection<VulnerableComponentView> vulnerableComponentViews, LinkableItem policyNameItem, LinkableItem policySeverity,
+        String componentName, String componentVersionName, String projectVersionUrl, Long notificationId, BlackDuckService blackDuckService, BlackDuckResponseCache blackDuckResponseCache) {
+        Map<String, VulnerabilityView> vulnerabilityViews = vulnerabilityUtil.createVulnerabilityViewMap(blackDuckService, vulnerableComponentViews);
+        List<VulnerabilityWithRemediationView> notificationVulnerabilities = vulnerableComponentViews.stream()
+                                                                                 .map(VulnerableComponentView::getVulnerabilityWithRemediation)
+                                                                                 .sorted(Comparator.comparing(VulnerabilityWithRemediationView::getSeverity))
+                                                                                 .collect(Collectors.toList());
+        SetMap<LinkableItem, LinkableItem> severityToVulns = SetMap.createLinked();
+        for (VulnerabilityWithRemediationView vulnerabilityView : notificationVulnerabilities) {
+            // TODO to get the URLS for vulnerabilities we would want to traverse the vulnerabilities link
+            String vulnerabilityId = vulnerabilityView.getVulnerabilityName();
+            String vulnerabilityUrl = null;
+            if (vulnerabilityViews.containsKey(vulnerabilityId)) {
+                vulnerabilityUrl = vulnerabilityViews.get(vulnerabilityId).getHref().orElse(null);
+            }
+
+            LinkableItem vulnerabilityIdItem = new LinkableItem(MessageBuilderConstants.LABEL_VULNERABILITIES, vulnerabilityId, vulnerabilityUrl);
+            vulnerabilityIdItem.setCollapsible(true);
+            LinkableItem severityItem = new LinkableItem(MessageBuilderConstants.LABEL_VULNERABILITY_SEVERITY, blackDuckResponseCache.getSeverity(vulnerabilityUrl));
+
+            severityToVulns.add(severityItem, vulnerabilityIdItem);
+        }
+
+        List<ComponentItem> vulnerabilityItems = new ArrayList<>();
+        for (Map.Entry<LinkableItem, Set<LinkableItem>> groupedVulnEntries : severityToVulns.entrySet()) {
+            String severityValue = groupedVulnEntries.getKey().getValue();
+            ComponentItemPriority priority = ComponentItemPriority.findPriority(severityValue);
+
+            List<LinkableItem> vulnAttributes = groupedVulnEntries.getValue()
+                                                    .stream()
+                                                    .map(vulnItem -> vulnerabilityUtil.createVulnerabilityAttributeItem(severityValue, vulnItem))
+                                                    .collect(Collectors.toList());
+            ComponentItem.Builder builder = new ComponentItem.Builder()
+                                                .applyCategory(MessageBuilderConstants.CATEGORY_TYPE_POLICY)
+                                                .applyOperation(ItemOperation.INFO)
+                                                .applyPriority(priority)
+                                                .applyCategoryItem(policyNameItem)
+                                                .applyCategoryGroupingAttribute(policySeverity)
+                                                .applyCollapseOnCategory(false)
+                                                .applyAllComponentAttributes(vulnAttributes)
+                                                .applyNotificationId(notificationId);
+            componentBuilderUtil.applyComponentInformation(builder, blackDuckResponseCache, componentName, componentVersionName, projectVersionUrl);
+            try {
+                vulnerabilityItems.add(builder.build());
+            } catch (AlertException ex) {
+                logger.info("Error building policy bom edit component for notification {}, operation {}, component {}, component version {}", notificationId, ItemOperation.UPDATE, componentName, componentVersionName);
+                logger.error("Error building policy bom edit component cause ", ex);
+            }
+        }
+
+        return vulnerabilityItems;
+    }
+}
