+++ src/main/java/com/synopsys/integration/alert/provider/blackduck/collector/builder/policy/PolicyViolationMessageBuilder.java
@@ -23,8 +23,6 @@
 package com.synopsys.integration.alert.provider.blackduck.collector.builder.policy;
 
 import java.util.ArrayList;
-import java.util.Collection;
-import java.util.Comparator;
 import java.util.Date;
 import java.util.LinkedList;
 import java.util.List;
@@ -34,7 +32,6 @@ import java.util.Set;
 import java.util.function.Function;
 import java.util.stream.Collectors;
 
-import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -54,14 +51,10 @@ import com.synopsys.integration.alert.provider.blackduck.collector.builder.util.
 import com.synopsys.integration.alert.provider.blackduck.collector.builder.util.PolicyPriorityUtil;
 import com.synopsys.integration.alert.provider.blackduck.collector.builder.util.VulnerabilityUtil;
 import com.synopsys.integration.alert.provider.blackduck.collector.util.BlackDuckResponseCache;
-import com.synopsys.integration.blackduck.api.generated.component.PolicyRuleExpressionSetView;
-import com.synopsys.integration.blackduck.api.generated.component.PolicyRuleExpressionView;
 import com.synopsys.integration.blackduck.api.generated.enumeration.NotificationType;
 import com.synopsys.integration.blackduck.api.generated.view.ComponentVersionView;
 import com.synopsys.integration.blackduck.api.generated.view.PolicyRuleView;
 import com.synopsys.integration.blackduck.api.generated.view.VersionBomComponentView;
-import com.synopsys.integration.blackduck.api.generated.view.VulnerabilityView;
-import com.synopsys.integration.blackduck.api.generated.view.VulnerabilityWithRemediationView;
 import com.synopsys.integration.blackduck.api.generated.view.VulnerableComponentView;
 import com.synopsys.integration.blackduck.api.manual.component.ComponentVersionStatus;
 import com.synopsys.integration.blackduck.api.manual.component.PolicyInfo;
@@ -81,12 +74,14 @@ public class PolicyViolationMessageBuilder implements BlackDuckMessageBuilder<Ru
     private PolicyPriorityUtil policyPriorityUtil;
     private VulnerabilityUtil vulnerabilityUtil;
     private ComponentBuilderUtil componentBuilderUtil;
+    private PolicyCommonBuilder policyCommonBuilder;
 
     @Autowired
-    public PolicyViolationMessageBuilder(PolicyPriorityUtil policyPriorityUtil, VulnerabilityUtil vulnerabilityUtil, ComponentBuilderUtil componentBuilderUtil) {
+    public PolicyViolationMessageBuilder(PolicyPriorityUtil policyPriorityUtil, VulnerabilityUtil vulnerabilityUtil, ComponentBuilderUtil componentBuilderUtil, PolicyCommonBuilder policyCommonBuilder) {
         this.policyPriorityUtil = policyPriorityUtil;
         this.vulnerabilityUtil = vulnerabilityUtil;
         this.componentBuilderUtil = componentBuilderUtil;
+        this.policyCommonBuilder = policyCommonBuilder;
     }
 
     @Override
@@ -111,7 +106,7 @@ public class PolicyViolationMessageBuilder implements BlackDuckMessageBuilder<Ru
                                                                               .applySubTopic(MessageBuilderConstants.LABEL_PROJECT_VERSION_NAME, violationContent.getProjectVersionName(), violationContent.getProjectVersion())
                                                                               .applyProviderCreationTime(providerCreationDate);
             Map<String, PolicyInfo> policyUrlToInfoMap = violationContent.getPolicyInfos().stream().collect(Collectors.toMap(PolicyInfo::getPolicy, Function.identity()));
-            SetMap<ComponentVersionStatus, PolicyInfo> componentPolicies = createComponentToPolicyMapping(violationContent.getComponentVersionStatuses(), policyUrlToInfoMap);
+            SetMap<ComponentVersionStatus, PolicyInfo> componentPolicies = policyCommonBuilder.createComponentToPolicyMapping(violationContent.getComponentVersionStatuses(), policyUrlToInfoMap);
             List<ComponentItem> items = new LinkedList<>();
             for (Map.Entry<ComponentVersionStatus, Set<PolicyInfo>> componentToPolicyEntry : componentPolicies.entrySet()) {
                 ComponentVersionStatus componentVersionStatus = componentToPolicyEntry.getKey();
@@ -142,8 +137,8 @@ public class PolicyViolationMessageBuilder implements BlackDuckMessageBuilder<Ru
             Optional<VersionBomComponentView> optionalBomComponent = blackDuckResponseCache.getBomComponentView(bomComponentUrl);
 
             List<LinkableItem> policyAttributes = new ArrayList<>();
-            LinkableItem policyNameItem = createPolicyNameItem(policyInfo);
-            LinkableItem nullablePolicySeverityItem = createPolicySeverityItem(policyInfo).orElse(null);
+            LinkableItem policyNameItem = componentBuilderUtil.createPolicyNameItem(policyInfo);
+            LinkableItem nullablePolicySeverityItem = componentBuilderUtil.createPolicySeverityItem(policyInfo).orElse(null);
             optionalBomComponent.ifPresent(bomComponent -> {
                 policyAttributes.addAll(componentBuilderUtil.getLicenseLinkableItems(bomComponent));
                 policyAttributes.addAll(componentBuilderUtil.getUsageLinkableItems(bomComponent));
@@ -166,8 +161,10 @@ public class PolicyViolationMessageBuilder implements BlackDuckMessageBuilder<Ru
                 logger.info("Error building policy component for notification {}, operation {}, component {}, component version {}", notificationId, operation, componentName, componentVersionName);
                 logger.error("Error building policy component cause ", ex);
             }
-            Optional<PolicyRuleView> optionalPolicyRule = getPolicyRule(blackDuckResponseCache, policyInfo);
-            if (optionalPolicyRule.isPresent() && hasVulnerabilityRule(optionalPolicyRule.get())) {
+            //        policyCommonBuilder.retrievePolicyItems(blackDuckResponseCache, componentVersionStatus.getComponentName(), componentVersionStatus.getComponentVersionName(), policies, notificationId, operation, projectVersionUrl,
+            //            componentVersionStatus.getBomComponent(), List.of());
+            Optional<PolicyRuleView> optionalPolicyRule = blackDuckResponseCache.getPolicyRule(blackDuckResponseCache, policyInfo);
+            if (optionalPolicyRule.isPresent() && policyCommonBuilder.hasVulnerabilityRule(optionalPolicyRule.get())) {
                 if (optionalBomComponent.isPresent()) {
                     List<ComponentItem> vulnerabilityPolicyItems =
                         createVulnerabilityPolicyItems(blackDuckResponseCache, blackDuckService, componentService, optionalBomComponent.get(), policyNameItem, nullablePolicySeverityItem, componentName, componentVersionName,
@@ -177,61 +174,6 @@ public class PolicyViolationMessageBuilder implements BlackDuckMessageBuilder<Ru
             }
         }
         return componentItems;
-    }
-
-    protected LinkableItem createPolicyNameItem(PolicyInfo policyInfo) {
-        String policyName = policyInfo.getPolicyName();
-        return new LinkableItem(MessageBuilderConstants.LABEL_POLICY_NAME, policyName);
-    }
-
-    protected Optional<LinkableItem> createPolicySeverityItem(PolicyInfo policyInfo) {
-        String severity = policyInfo.getSeverity();
-        if (StringUtils.isNotBlank(severity)) {
-            LinkableItem severityItem = new LinkableItem(MessageBuilderConstants.LABEL_POLICY_SEVERITY_NAME, severity);
-            return Optional.of(severityItem);
-        }
-        return Optional.empty();
-    }
-
-    public boolean hasVulnerabilityRule(PolicyRuleView policyRule) {
-        PolicyRuleExpressionSetView expression = policyRule.getExpression();
-        List<PolicyRuleExpressionView> expressions = expression.getExpressions();
-        for (PolicyRuleExpressionView expressionView : expressions) {
-            if (expressionView.getName().toLowerCase().contains(MessageBuilderConstants.VULNERABILITY_CHECK_TEXT)) {
-                return true;
-            }
-        }
-        return false;
-    }
-
-    public Optional<PolicyRuleView> getPolicyRule(BlackDuckResponseCache blackDuckResponseCache, PolicyInfo policyInfo) {
-        try {
-            String policyUrl = policyInfo.getPolicy();
-            if (StringUtils.isNotBlank(policyUrl)) {
-                return blackDuckResponseCache.getItem(PolicyRuleView.class, policyUrl);
-            }
-        } catch (Exception e) {
-            logger.debug("Unable to get policy rule: {}", policyInfo.getPolicyName());
-            logger.debug("Cause:", e);
-        }
-        return Optional.empty();
-    }
-
-    private SetMap<ComponentVersionStatus, PolicyInfo> createComponentToPolicyMapping(
-        Collection<ComponentVersionStatus> componentVersionStatuses, Map<String, PolicyInfo> policyItems) {
-        SetMap<ComponentVersionStatus, PolicyInfo> componentToPolicyMapping = SetMap.createDefault();
-        for (ComponentVersionStatus componentVersionStatus : componentVersionStatuses) {
-            Set<PolicyInfo> componentPolicies = getPoliciesForComponent(componentVersionStatus, policyItems);
-            componentToPolicyMapping.addAll(componentVersionStatus, componentPolicies);
-        }
-        return componentToPolicyMapping;
-    }
-
-    private Set<PolicyInfo> getPoliciesForComponent(ComponentVersionStatus componentVersionStatus, Map<String, PolicyInfo> policyItems) {
-        return componentVersionStatus.getPolicies().stream()
-                   .filter(policyItems::containsKey)
-                   .map(policyItems::get)
-                   .collect(Collectors.toSet());
     }
 
     private List<ComponentItem> createVulnerabilityPolicyItems(BlackDuckResponseCache blackDuckResponseCache, BlackDuckService blackDuckService, ComponentService componentService, VersionBomComponentView bomComponent,
@@ -244,8 +186,9 @@ public class PolicyViolationMessageBuilder implements BlackDuckMessageBuilder<Ru
                 ProjectVersionWrapper projectVersionWrapper = optionalProjectVersionWrapper.get();
                 String projectVersionUrl = projectVersionWrapper.getProjectVersionView().getHref().orElse(null);
                 List<VulnerableComponentView> vulnerableComponentViews = vulnerabilityUtil.getVulnerableComponentViews(blackDuckService, projectVersionWrapper, bomComponent);
-                List<ComponentItem> vulnerabilityComponentItems =
-                    createVulnerabilityPolicyComponentItems(vulnerableComponentViews, policyNameItem, policySeverity, componentName, componentVersionName, projectVersionUrl, notificationId, blackDuckService, blackDuckResponseCache);
+                List<ComponentItem> vulnerabilityComponentItems = policyCommonBuilder
+                                                                      .createVulnerabilityPolicyComponentItems(vulnerableComponentViews, policyNameItem, policySeverity, componentName, componentVersionName, projectVersionUrl, notificationId,
+                                                                          blackDuckService, blackDuckResponseCache);
                 vulnerabilityPolicyItems.addAll(vulnerabilityComponentItems);
 
                 // TODO: remove the orElse null.
@@ -264,8 +207,7 @@ public class PolicyViolationMessageBuilder implements BlackDuckMessageBuilder<Ru
     }
 
     protected Optional<ComponentItem> createRemediationComponentItem(BlackDuckResponseCache blackDuckResponseCache, String categoryType, ComponentService componentService, ComponentVersionView componentVersionView, String componentName,
-        String componentVersionName, String projectVersionURL,
-        LinkableItem categoryItem, LinkableItem categoryGrouping, boolean collapseOnCategory, Long notificationId) {
+        String componentVersionName, String projectVersionURL, LinkableItem categoryItem, LinkableItem categoryGrouping, boolean collapseOnCategory, Long notificationId) {
         try {
             List<LinkableItem> remediationItems = vulnerabilityUtil.getRemediationItems(componentService, componentVersionView);
             if (!remediationItems.isEmpty()) {
@@ -286,59 +228,6 @@ public class PolicyViolationMessageBuilder implements BlackDuckMessageBuilder<Ru
             logger.debug("Could not create remediation component", e);
         }
         return Optional.empty();
-    }
-
-    protected List<ComponentItem> createVulnerabilityPolicyComponentItems(Collection<VulnerableComponentView> vulnerableComponentViews, LinkableItem policyNameItem, LinkableItem policySeverity,
-        String componentName, String componentVersionName, String projectVersionUrl, Long notificationId, BlackDuckService blackDuckService, BlackDuckResponseCache blackDuckResponseCache) {
-        Map<String, VulnerabilityView> vulnerabilityViews = vulnerabilityUtil.createVulnerabilityViewMap(blackDuckService, vulnerableComponentViews);
-        List<VulnerabilityWithRemediationView> notificationVulnerabilities = vulnerableComponentViews.stream()
-                                                                                 .map(VulnerableComponentView::getVulnerabilityWithRemediation)
-                                                                                 .sorted(Comparator.comparing(VulnerabilityWithRemediationView::getSeverity))
-                                                                                 .collect(Collectors.toList());
-        SetMap<LinkableItem, LinkableItem> severityToVulns = SetMap.createLinked();
-        for (VulnerabilityWithRemediationView vulnerabilityView : notificationVulnerabilities) {
-            // TODO to get the URLS for vulnerabilities we would want to traverse the vulnerabilities link
-            String vulnerabilityId = vulnerabilityView.getVulnerabilityName();
-            String vulnerabilityUrl = null;
-            if (vulnerabilityViews.containsKey(vulnerabilityId)) {
-                vulnerabilityUrl = vulnerabilityViews.get(vulnerabilityId).getHref().orElse(null);
-            }
-
-            LinkableItem vulnerabilityIdItem = new LinkableItem(MessageBuilderConstants.LABEL_VULNERABILITIES, vulnerabilityId, vulnerabilityUrl);
-            vulnerabilityIdItem.setCollapsible(true);
-            LinkableItem severityItem = new LinkableItem(MessageBuilderConstants.LABEL_VULNERABILITY_SEVERITY, blackDuckResponseCache.getSeverity(vulnerabilityUrl));
-
-            severityToVulns.add(severityItem, vulnerabilityIdItem);
-        }
-
-        List<ComponentItem> vulnerabilityItems = new ArrayList<>();
-        for (Map.Entry<LinkableItem, Set<LinkableItem>> groupedVulnEntries : severityToVulns.entrySet()) {
-            String severityValue = groupedVulnEntries.getKey().getValue();
-            ComponentItemPriority priority = ComponentItemPriority.findPriority(severityValue);
-
-            List<LinkableItem> vulnAttributes = groupedVulnEntries.getValue()
-                                                    .stream()
-                                                    .map(vulnItem -> vulnerabilityUtil.createVulnerabilityAttributeItem(severityValue, vulnItem))
-                                                    .collect(Collectors.toList());
-            ComponentItem.Builder builder = new ComponentItem.Builder()
-                                                .applyCategory(MessageBuilderConstants.CATEGORY_TYPE_POLICY)
-                                                .applyOperation(ItemOperation.INFO)
-                                                .applyPriority(priority)
-                                                .applyCategoryItem(policyNameItem)
-                                                .applyCategoryGroupingAttribute(policySeverity)
-                                                .applyCollapseOnCategory(false)
-                                                .applyAllComponentAttributes(vulnAttributes)
-                                                .applyNotificationId(notificationId);
-            componentBuilderUtil.applyComponentInformation(builder, blackDuckResponseCache, componentName, componentVersionName, projectVersionUrl);
-            try {
-                vulnerabilityItems.add(builder.build());
-            } catch (AlertException ex) {
-                logger.info("Error building policy bom edit component for notification {}, operation {}, component {}, component version {}", notificationId, ItemOperation.UPDATE, componentName, componentVersionName);
-                logger.error("Error building policy bom edit component cause ", ex);
-            }
-        }
-
-        return vulnerabilityItems;
     }
 
 }
