+++ src/main/java/com/synopsys/integration/alert/provider/blackduck/collector/builder/util/VulnerabilityUtil.java
@@ -43,9 +43,12 @@ import org.springframework.stereotype.Component;
 import com.synopsys.integration.alert.common.message.model.LinkableItem;
 import com.synopsys.integration.alert.provider.blackduck.collector.builder.MessageBuilderConstants;
 import com.synopsys.integration.blackduck.api.generated.component.RemediatingVersionView;
+import com.synopsys.integration.blackduck.api.generated.component.RiskCountView;
+import com.synopsys.integration.blackduck.api.generated.enumeration.RiskCountType;
 import com.synopsys.integration.blackduck.api.generated.response.RemediationOptionsView;
 import com.synopsys.integration.blackduck.api.generated.view.ComponentVersionView;
 import com.synopsys.integration.blackduck.api.generated.view.ProjectVersionView;
+import com.synopsys.integration.blackduck.api.generated.view.RiskProfileView;
 import com.synopsys.integration.blackduck.api.generated.view.VersionBomComponentView;
 import com.synopsys.integration.blackduck.api.generated.view.VulnerabilityView;
 import com.synopsys.integration.blackduck.api.generated.view.VulnerabilityWithRemediationView;
@@ -135,5 +138,26 @@ public class VulnerabilityUtil {
             stringBuilder.append(")");
         }
         return stringBuilder.toString();
+    }
+
+    public boolean doesSecurityRiskProfileHaveVulnerabilities(final RiskProfileView securityRiskProfile) {
+        logger.debug("Checking if the component still has vulnerabilities...");
+        int vulnerabilitiesCount = getSumOfRiskCounts(securityRiskProfile.getCounts());
+        logger.debug("Number of vulnerabilities found: " + vulnerabilitiesCount);
+        if (vulnerabilitiesCount > 0) {
+            logger.debug("This component still has vulnerabilities");
+            return true;
+        }
+        return false;
+    }
+
+    private int getSumOfRiskCounts(List<RiskCountView> vulnerabilityCounts) {
+        int count = 0;
+        for (RiskCountView riskCount : vulnerabilityCounts) {
+            if (!RiskCountType.OK.equals(riskCount.getCountType())) {
+                count += riskCount.getCount();
+            }
+        }
+        return count;
     }
 }
