+++ src/test/java/com/synopsys/integration/alert/provider/blackduck/collector/PolicyViolationMessageBuilderTest.java
@@ -18,6 +18,7 @@ import com.synopsys.integration.alert.common.message.model.ProviderMessageConten
 import com.synopsys.integration.alert.common.persistence.model.ConfigurationJobModel;
 import com.synopsys.integration.alert.provider.blackduck.collector.builder.BlackDuckMessageBuilder;
 import com.synopsys.integration.alert.provider.blackduck.collector.builder.policy.PolicyClearedMessageBuilder;
+import com.synopsys.integration.alert.provider.blackduck.collector.builder.policy.PolicyCommonBuilder;
 import com.synopsys.integration.alert.provider.blackduck.collector.builder.policy.PolicyViolationMessageBuilder;
 import com.synopsys.integration.alert.provider.blackduck.collector.builder.util.ComponentBuilderUtil;
 import com.synopsys.integration.alert.provider.blackduck.collector.builder.util.PolicyPriorityUtil;
@@ -36,7 +37,9 @@ public class PolicyViolationMessageBuilderTest {
     public void insertRuleViolationClearedNotificationTest() throws Exception {
         PolicyPriorityUtil policyPriorityUtil = new PolicyPriorityUtil();
         ComponentBuilderUtil componentBuilderUtil = new ComponentBuilderUtil();
-        PolicyClearedMessageBuilder policyViolationClearedMessageBuilder = new PolicyClearedMessageBuilder(policyPriorityUtil, componentBuilderUtil);
+        VulnerabilityUtil vulnerabilityUtil = new VulnerabilityUtil();
+        PolicyCommonBuilder policyCommonBuilder = new PolicyCommonBuilder(vulnerabilityUtil, componentBuilderUtil, policyPriorityUtil);
+        PolicyClearedMessageBuilder policyViolationClearedMessageBuilder = new PolicyClearedMessageBuilder(policyPriorityUtil, componentBuilderUtil, policyCommonBuilder);
         runSingleTest(policyViolationClearedMessageBuilder, TestConstants.POLICY_CLEARED_NOTIFICATION_JSON_PATH, NotificationType.RULE_VIOLATION_CLEARED);
     }
 
@@ -45,7 +48,8 @@ public class PolicyViolationMessageBuilderTest {
         PolicyPriorityUtil policyPriorityUtil = new PolicyPriorityUtil();
         ComponentBuilderUtil componentBuilderUtil = new ComponentBuilderUtil();
         VulnerabilityUtil vulnUtil = Mockito.mock(VulnerabilityUtil.class);
-        PolicyViolationMessageBuilder policyViolationMessageBuilder = new PolicyViolationMessageBuilder(policyPriorityUtil, vulnUtil, componentBuilderUtil);
+        PolicyCommonBuilder policyCommonBuilder = new PolicyCommonBuilder(vulnUtil, componentBuilderUtil, policyPriorityUtil);
+        PolicyViolationMessageBuilder policyViolationMessageBuilder = new PolicyViolationMessageBuilder(policyPriorityUtil, vulnUtil, componentBuilderUtil, policyCommonBuilder);
         runSingleTest(policyViolationMessageBuilder, TestConstants.POLICY_CLEARED_NOTIFICATION_JSON_PATH, NotificationType.RULE_VIOLATION);
     }
